+++ premium/web-frontend/modules/baserow_premium/components/views/kanban/KanbanViewStackedBy.vue
@@ -29,6 +29,7 @@
         ref="createFieldContext"
         :table="table"
         :forced-type="forcedFieldType"
+        @field-created="$event.callback()"
       ></CreateFieldContext>
     </div>
   </div>